+++ src/main/java/com/github/craxlor/discordbot/command/module/autoroom/command/AutoroomCollection.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import com.github.craxlor.discordbot.command.module.autoroom.command.slash.Customize;
 import com.github.craxlor.discordbot.command.module.autoroom.command.slash.Setup;
 import com.github.craxlor.discordbot.command.slash.SlashCommand;
-import com.github.craxlor.discordbot.manager.commandlist.CommandCollection;
+import com.github.craxlor.discordbot.util.core.commandlist.CommandCollection;
 
 public class AutoroomCollection extends ArrayList<SlashCommand> implements CommandCollection {
