+++ src/main/java/com/github/craxlor/discordbot/command/module/core/command/slash/Shutdown.java
@@ -4,7 +4,7 @@ import javax.annotation.Nonnull;
 
 import com.github.craxlor.discordbot.command.slash.SCDev;
 import com.github.craxlor.discordbot.database.Database;
-import com.github.craxlor.discordbot.manager.GuildManager;
+import com.github.craxlor.discordbot.util.core.GuildManager;
 import com.github.craxlor.discordbot.util.reply.LogHelper;
 import com.github.craxlor.discordbot.util.reply.Reply;
 import com.github.craxlor.discordbot.util.reply.Status;
