+++ src/main/java/com/github/craxlor/discordbot/command/module/music/command/MusicCollection.java
@@ -11,7 +11,7 @@ import com.github.craxlor.discordbot.command.module.music.command.slash.Resume;
 import com.github.craxlor.discordbot.command.module.music.command.slash.Skip;
 import com.github.craxlor.discordbot.command.module.music.command.slash.Song;
 import com.github.craxlor.discordbot.command.slash.SlashCommand;
-import com.github.craxlor.discordbot.manager.commandlist.CommandCollection;
+import com.github.craxlor.discordbot.util.core.commandlist.CommandCollection;
 
 public class MusicCollection extends ArrayList<SlashCommand> implements CommandCollection {
