+++ src/main/java/com/github/craxlor/discordbot/command/module/music/command/slash/Play.java
@@ -7,8 +7,8 @@ import javax.annotation.Nonnull;
 import com.github.craxlor.discordbot.command.slash.SCMusic;
 import com.github.craxlor.discordbot.database.Database;
 import com.github.craxlor.discordbot.database.element.YouTubeSearch;
-import com.github.craxlor.discordbot.manager.GuildManager;
-import com.github.craxlor.discordbot.manager.MusicManager;
+import com.github.craxlor.discordbot.util.core.GuildManager;
+import com.github.craxlor.discordbot.util.music.MusicManager;
 import com.github.craxlor.discordbot.util.music.SpotifyHelper;
 import com.github.craxlor.discordbot.util.music.YouTubeHelper;
 import com.github.craxlor.discordbot.util.reply.Reply;
