+++ src/main/java/com/github/craxlor/discordbot/command/slash/SCMusic.java
@@ -5,7 +5,7 @@ import org.slf4j.Logger;
 import com.github.craxlor.discordbot.Properties;
 import com.github.craxlor.discordbot.command.module.music.command.slash.Play;
 import com.github.craxlor.discordbot.database.Database;
-import com.github.craxlor.discordbot.manager.GuildManager;
+import com.github.craxlor.discordbot.util.core.GuildManager;
 
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
