+++ src/main/java/com/github/craxlor/discordbot/database/element/DiscordServer.java
@@ -86,7 +86,8 @@ public class DiscordServer {
     public String getModules() {
         return modules;
     }
-
+    
+    @Nullable
     public String getColorHex() {
         return colorHex;
     }
